split state into files
