+++ src/app/pages/regional/regional-initial-state/regional-charts.js
@@ -1,0 +1,31 @@
+export default {
+  localProtectedSpecies: {
+    data: [
+      {
+        subject: 'BIRDS',
+        percent: 96
+      },
+      {
+        subject: 'MAMMALS',
+        percent: 88
+      },
+      {
+        subject: 'AMPHIBIANS',
+        percent: 56
+      },
+      {
+        subject: 'CACTI',
+        percent: 100
+      },
+      {
+        subject: 'TURTLES',
+        percent: 81
+      },
+      {
+        subject: 'CONIFERS',
+        percent: 71
+      }
+    ],
+    dimensions: [{ key: 'percent' }]
+  }
+}
