+++ src/app/pages/regional/regional-initial-state/regional-initial-state.js
@@ -1,0 +1,13 @@
+import legend from './regional-legends'
+import birds from './regional-birds'
+import layers from './regional-layers'
+import sections from './regional-sections'
+import charts from './regional-charts'
+
+export default {
+  ...charts,
+  layers,
+  sections,
+  legend,
+  birds
+}
